choose_features_clustering

choose_features_clustering

choose_features_clustering(
  heteroplasmy_matrix,
  allele_matrix,
  cluster,
  top_pos,
  deepSplit_param,
  minClusterSize_param,
  min_value_vector,
  threshold = 0.2,
  index,
  max_frac = 0.7
)

Arguments

heteroplasmy_matrix

Third element returned by get_heteroplasmy.

allele_matrix

Fourth element returned by get_heteroplasmy.

cluster

Vector specifying a partition of the samples.

top_pos

Numeric value. Number of bases sorted with decreasing values of distance variance (see section Details below) among samples. If relevant_bases=NULL, then the bases for performing hierarchical clustering are the ones whose relative variance (variance of the base divided sum of variance among top_pos bases) is above min_value.

deepSplit_param

Integer value between 0 and 4 for the deepSplit parameter of the function cutreeHybrid. See section Details below.

minClusterSize_param

Integer value specifying the minClusterSize parameter of the function cutreeHybrid. See section Details below.

min_value_vector

Numeric vector. For each value in the vector, the function clustering_angular_distance is run with parameter min_value equal to one element of the vector min_value_vector.

threshold

Numeric value. If a base has heteroplasmy greater or equal to threshold in more than max_frac of cells, then the base is not considered for down stream analysis.

index

Fifth element returned by get_heteroplasmy.

max_frac

Numeric value.If a base has heteroplasmy greater or equal to threshold in more than max_frac of cells, then the base is not considered for down stream analysis.

Value

Clustree plot.

See also

Author

Gabriele Lubatti gabriele.lubatti@helmholtz-muenchen.de