get_wilcox_test

get_wilcox_test

get_wilcox_test(heteroplasmy_matrix, cluster, label_1, label_2, index = NULL)

Arguments

heteroplasmy_matrix

Third element returned by get_heteroplasmy.

cluster

Vector specifying a partition of the samples.

label_1

Character name of a first label included in cluster. It denotes the first group used for the Wilcoxon test

label_2

Character name of a second label included in cluster and different from label_1. it denotes the second group used for the Wilcoxon test.

index

Fifth element returned by get_heteroplasmy.

Value

It returns a vector of length equal to n_row in matrix. Each element stands for a base and it contains the adjusted p-value (FDR), obtained in unpaired two-samples Wilcoxon test from the comparison of the heteroplasmy between the label_1 and label_2 group.

See also

Author

Gabriele Lubatti gabriele.lubatti@helmholtz-muenchen.de